# backend/config/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alice Voice Backend API",
    "description": "REST and socket contract owned by the backend. Frontend copies DTOs from this file.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local daemon"
    }
  ],
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Reply of `POST /api/v1/chat`.",
        "description": "The daemon handles the message while the request runs: it resolves the\nintent, runs its command, and answers. Every stage streams on the\nsocket stream, so the interface shows the turn while it runs.\n\nThe daemon stores the turn in a conversation when the queue is on. It\nanswers without a store when the queue is off.",
        "operationId": "chat_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Handled turn",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatReplyDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Conversation not found"
          },
          "413": {
            "description": "Message too long"
          }
        }
      }
    },
...
    "/api/v1/scripts/{id}": {
      "get": {
        "tags": [
          "scripts"
        ],
        "summary": "Reply of `GET /api/v1/scripts/{id}`.",
        "operationId": "get_script_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Script id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stored script",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              }
            }
          },
          "404": {
            "description": "Script not found"
          }
        }
      }
    },
    "/api/v1/scripts/{id}/approve": {
      "post": {
        "tags": [
          "scripts"
        ],
        "summary": "Reply of `POST /api/v1/scripts/{id}/approve`.",
        "description": "The daemon claims the pending script and, when the claim belongs to\nthis call, runs it. A second approval of the same script reports the\nstored script and runs nothing.",
        "operationId": "approve_script_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Script id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stored script after the run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              }
            }
          },
          "404": {
            "description": "Script not found"
          }
        }
      }
    },
    "/api/v1/scripts/{id}/deny": {
      "post": {
        "tags": [
          "scripts"
        ],
        "summary": "Reply of `POST /api/v1/scripts/{id}/deny`.",
        "description": "A denied script stays in the store with its decision, so the surface\nstill reads it and the daemon never runs it.",
        "operationId": "deny_script_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Script id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stored script after the decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              }
            }
          },
          "404": {
            "description": "Script not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatMessageDto": {
        "type": "object",
        "description": "One stored turn of a conversation.",
        "required": [
          "id",
          "role",
          "text",
          "createdAt"
        ],
        "properties": {
          "confidence": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Resolver confidence between 0 and 1, or null."
          },
          "conversationId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Conversation the message belongs to, or null when the queue is off."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Time the daemon stored the message, in ISO 8601."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable message identifier."
          },
          "intentId": {
            "type": [
              "string",
              "null"
            ],
            "description": "Resolved intent identifier, or null when none."
          },
          "intentName": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name of the resolved intent, or null when none."
          },
          "memory": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MessageMemoryDto",
                "description": "\nWhat the turn taught the long term memory, or null when the memory\nlearned nothing from it. The memory reads a turn after the daemon\nanswered it, so the field is filled when a reader reads the message\nback rather than when the turn is stored."
              }
            ]
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MessageMetaDto",
                "description": "How the daemon read the turn, or null when it read nothing."
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/ChatRoleDto",
            "description": "Speaker."
          },
          "text": {
            "type": "string",
            "description": "Message text."
          }
        }
      },
      "ChatReplyDto": {
        "type": "object",
        "description": "Reply of `POST /api/v1/chat`.",
        "required": [
          "stored",
          "user",
          "reply"
        ],
        "properties": {
          "conversation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConversationDto",
                "description": "The conversation of the turn, or null when the queue is off."
              }
            ]
          },
          "reply": {
            "$ref": "#/components/schemas/ChatMessageDto",
            "description": "Daemon reply."
          },
          "stored": {
            "type": "boolean",
            "description": "True when the daemon stored the turn, false when the queue is off."
          },
          "user": {
            "$ref": "#/components/schemas/ChatMessageDto",
            "description": "Stored user message."
          }
        }
      },
      "ChatRequestDto": {
        "type": "object",
        "description": "Body of `POST /api/v1/chat`.",
        "required": [
          "text"
        ],
        "properties": {
          "context": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether the daemon reads the earlier turns of the conversation as\nthe context of this message. Null reads them.\n\nA message sent with this false is read on its own, so the router\nholds it against the catalog alone and a model answers without the\nturns before it. A follow-up whose meaning depends on the turn\nbefore it needs the context, so the default reads it."
          },
          "conversationId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Conversation to append to. Null starts a new conversation."
          },
          "text": {
            "type": "string",
            "description": "Text the user typed."
          }
        }
      },
      "ChatRoleDto": {
        "type": "string",
        "description": "Speaker of one chat message.",
        "enum": [
          "user",
          "assistant"
        ]
      },
...
      "ConversationDto": {
        "type": "object",
        "description": "One conversation of the daemon. The conversation owns its message history.",
        "required": [
          "id",
          "title",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Time the daemon created the conversation, in ISO 8601."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable conversation identifier."
          },
          "title": {
            "type": "string",
            "description": "Title the daemon generated from the first message."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Time the daemon stored the last message, in ISO 8601."
          }
        }
      },
...
      "MemorySavedFactDto": {
        "type": "object",
        "description": "One fact that one turn taught the long term memory.",
        "required": [
          "concept",
          "relation",
          "value"
        ],
        "properties": {
          "concept": {
            "type": "string",
            "description": "The name of the concept the fact belongs to."
          },
          "relation": {
            "type": "string",
            "description": "The relation the fact names."
          },
          "value": {
            "type": "string",
            "description": "The value the fact carries."
          }
        }
      },
...
      "MessageMemoryDto": {
        "type": "object",
        "description": "What one turn taught the long term memory.\n\nThe memory learns from the turns that met no intent, and it writes them\nin the background. This is what a reader sees of that work: the facts\nthe turn taught, each with the concept it belongs to.",
        "required": [
          "facts"
        ],
        "properties": {
          "facts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemorySavedFactDto"
            },
            "description": "The facts the turn taught, in the order the memory wrote them."
          }
        }
      },
      "MessageMetaDto": {
        "type": "object",
        "description": "How the daemon read one handled turn.\n\nThe resolver owns two steps: it chooses the intent, and it reads the\nvalues of the entities of that intent. Either step may run on the\nllama.cpp server or on the built in GLiNER model, so a turn reports\nthe engine of each step. The rest of the metadata is what the daemon\nthen ran, so the transcript can explain a stored turn.",
        "properties": {
          "candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageCandidateDto"
            },
            "description": "The intents the retrieval pass kept for this turn, best first.\n\nThe short list is the evidence of the decision, so a user who\nreads `no intent matched` still reads what the catalog offered."
          },
          "command": {
            "type": [
              "string",
              "null"
            ],
            "description": "The command after the daemon put the values in, or null."
          },
          "durationMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "How long the command ran, in milliseconds, or null when none ran.",
            "minimum": 0
          },
          "entities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageEntityDto"
            },
            "description": "The entities of the intent with the values the daemon read."
          },
          "exitCode": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The exit code of the command, or null when none ran."
          },
          "intentEngine": {
            "type": [
              "string",
              "null"
            ],
            "description": "The engine that chose the intent, or null when none chose one."
          },
          "intentModel": {
            "type": [
              "string",
              "null"
            ],
            "description": "The model that engine ran, or null when the daemon cannot name it."
          },
          "memorySeed": {
            "type": [
              "string",
              "null"
            ],
            "description": "The concepts the daemon read from the long term memory for this\nturn, as plain text, or null when the memory carried none.\n\nThe memory belongs to the branch that no intent matches, so only a\nturn without an intent reports what the daemon read of it: the\nfacts the answer of the turn leaned on."
          },
          "resolveMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "How long the daemon needed to read the turn, in milliseconds, or\nnull when it did not measure the read.\n\nThe read runs from the queued message to the command, so a slow\nturn is told apart from a slow command: the duration above is the\ncommand alone and this is the resolver.",
            "minimum": 0
          },
          "route": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MessageRouteDto",
                "description": "How the layered router read the turn, stage by stage, or null when\nthe backend of the turn is not the router.\n\nThe stage above names the stage that decided, and this names every\nstage the turn passed with the reader it ran and the short list it\nread, so a stored turn can be read back the way the daemon read it."
              }
            ]
          },
          "scriptId": {
            "type": [
              "string",
              "null"
            ],
            "description": "The identifier of the script the model wrote for this turn, or\nnull when the model answered in words.\n\nThe store keeps the script, so a user can still decide about a\nturn that a restart or a closed socket left behind."
          },
          "stage": {
            "type": [
              "string",
              "null"
            ],
            "description": "The stage of the layered router that decided the turn, or null\nwhen the backend of the turn is not the router.\n\nA turn the deterministic pass proved reads `fast_path`, a turn the\nretrieval scores decided reads `retrieve`, and a turn a model\nchose from the short list reads `rerank`."
          },
          "valueEngine": {
            "type": [
              "string",
              "null"
            ],
            "description": "The engine that read the entity values, or null when none did."
          },
          "valueModel": {
            "type": [
              "string",
              "null"
            ],
            "description": "The model that engine ran, or null when the daemon cannot name it."
          }
        }
      },
...
      "ScriptDto": {
        "type": "object",
        "description": "One shell script that waits for the decision of the user.",
        "required": [
          "id",
          "requestText",
          "summary",
          "script",
          "destructiveness",
          "status",
          "createdAt"
        ],
        "properties": {
          "conversationId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Conversation the turn belongs to, or null when the queue is off."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Time the daemon stored the script, in ISO 8601."
          },
          "destructiveness": {
            "type": "integer",
            "format": "int32",
            "description": "How rough the script is on the machine, between 0 and 100.",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable identifier of the script."
          },
          "requestText": {
            "type": "string",
            "description": "The message the model wrote the script for."
          },
          "script": {
            "type": "string",
            "description": "The shell script."
          },
          "status": {
            "type": "string",
            "description": "The decision of the user: pending, approved, denied, or ran."
          },
          "summary": {
            "type": "string",
            "description": "One sentence about what the script does."
          }
        }
      },
...
    }
  }
}